requirements3.txt
opencv-python
qrcode
numpy
Pillow